Move classes into seperated files.
