+++ osu.Framework/Localisation/LocalisationEngine.cs
@@ -157,77 +157,5 @@ namespace osu.Framework.Localisation
                     formattableBindings.Remove(w);
             }
         }
-
-        /// <summary>
-        /// A Bindable string which takes a unicode and non-unicode (usually romanised) version of the contained text
-        /// and provides automatic switching behaviour should the user change their preference.
-        /// </summary>
-        public class UnicodeBindableString : Bindable<string>
-        {
-            public readonly string Unicode;
-            public readonly string NonUnicode;
-
-            public UnicodeBindableString(string unicode, string nonUnicode) : base(nonUnicode)
-            {
-                Unicode = unicode;
-                NonUnicode = nonUnicode;
-
-                if (Unicode == null)
-                    Unicode = NonUnicode;
-                if (NonUnicode == null)
-                    NonUnicode = Unicode;
-            }
-
-            public bool PreferUnicode
-            {
-                get { return Value == Unicode; }
-                set { Value = value ? Unicode : NonUnicode; }
-            }
-        }
-
-        /// <summary>
-        /// A Bindable string which stays up-to-date with the current locale choice for the specified key.
-        /// </summary>
-        public class LocalisedString : Bindable<string>
-        {
-            public readonly string Key;
-            public LocalisedString(string key)
-            {
-                Key = key;
-            }
-        }
-
-        /// <summary>
-        /// A bindable string constructed from <see cref="string.Format(string, object[])"/>.
-        /// </summary>
-        public class FormattableString : Bindable<string>
-        {
-            private string format;
-            private readonly object[] objects;
-            protected virtual string Format => format;
-            public void Update() => Value = string.Format(Format, objects);
-
-            public FormattableString(string format, object[] objects)
-            {
-                this.format = format;
-                this.objects = objects;
-                Update();
-            }
-        }
-
-        /// <summary>
-        /// A bindable string constructed from <see cref="string.Format(string, object[])"/>, and <see cref="VaraintFormattableString.Format"/> changable.
-        /// </summary>
-        public class VaraintFormattableString : FormattableString
-        {
-            private readonly Bindable<string> formatSource;
-            protected override string Format => formatSource.Value;
-
-            public VaraintFormattableString(Bindable<string> formatSource, object[] objects)
-                : base(null, objects)
-            {
-                this.formatSource = formatSource;
-            }
-        }
     }
 }
